revisi di insert dan update
