+++ .gitignore
@@ -23,3 +23,4 @@ go.work
 
 # Current Workspace
 vendor
+archs
