--- coba/testtable.sql
@@ -1,6 +1,0 @@
-create Table temp.tmp_test (
-    id varchar(10),
-    nama varchar(30),
-
-
-)
